Move shared non-entry point JS files into 'base/' dir
