+++ h/static/scripts/admin-site.js
@@ -1,9 +1,9 @@
 'use strict';
 
 // configure error reporting
-var settings = require('./settings')(document);
+var settings = require('./base/settings')(document);
 if (settings.raven) {
-  require('./raven').init(settings.raven);
+  require('./base/raven').init(settings.raven);
 }
 
 window.$ = window.jQuery = require('jquery');
