+++ h/static/scripts/base/settings.js
